postgresql

+++ db/migrate/20141220035215_devise_create_users.rb
@@ -3,7 +3,7 @@ class DeviseCreateUsers < ActiveRecord::Migration
     create_table(:users) do |t|
       t.string :name
       t.integer :year
-      t.integer :phone
+      t.float :phone
 
       ## Database authenticatable
       t.string :email,              null: false, default: ""
